Basic spiral
You can change the rate/ speed of growth by tweaking the update
function of theta and the radius. Next I am working on creating a
spiral spline
that pushes points into an array and then draws them with a thin line
then I’ll try to get the camera to move along the spline.